Added transform data analysis sales.
